requirements.txt
absl-py==0.8.1
astor==0.8.0
<<<<<<< HEAD
bleach==1.5.0
Click==7.0
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
html5lib==0.9999999
=======
Click==7.0
Flask==1.1.1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
>>>>>>> 9a0ee4467f15d6b9dccb067cdd74ece47191053a
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.3
<<<<<<< HEAD
opt-einsum==3.1.0
Pillow==6.2.1
pkg-resources==0.0.0
protobuf==3.10.0
six==1.12.0
tensorboard==1.7.0
tensorflow==1.7.1
tensorflow-estimator==2.0.1
=======
Pillow==6.2.1
protobuf==3.10.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.14.0
>>>>>>> 9a0ee4467f15d6b9dccb067cdd74ece47191053a
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
